requirements.txt

# General
future==0.16.0
termcolor==1.1.0
# Web
Jinja2==2.10
##  Mapping
# Basemap
Shapely==1.6.4.post1
Fiona==1.7.11.post1
descartes==1.1.0
-e git+git@github.com:matplotlib/basemap.git@b24b3ef8f7b92252ee713d5f3151c248dbee5c0e#egg=basemap
pyproj==1.9.5.1
lxml==4.2.0
pyshp==1.2.12
#PySal lite mapping
PySAL==1.14.3
# Ipython and Jupyer
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.4.0
jupyter-core==4.4.0
jupyter-highlight-selected-word==0.1.0
jupyter-latex-envs==1.4.4
jupyter-nbextensions-configurator==0.4.0
ipykernel==4.8.2
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==7.1.2
qtconsole==4.3.1
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.0
appnope==0.1.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
bleach==2.1.3
configparser==3.5.0
decorator==4.2.1
entrypoints==0.2.3
functools32==3.2.3.post2
futures==3.2.0
html5lib==1.0.1
jsonschema==2.6.0
MarkupSafe==1.0
mistune==0.8.3
pathlib2==2.3.0
pexpect==4.4.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.7.0
pyzmq==17.0.0
scandir==1.7
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
terminado==0.8.1
testpath==0.3.1
tornado==5.0.1
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.1.4
# Numeric analysis
numpy==1.14.2
pandas==0.22.0
pytz==2018.3
scipy==1.0.0
# Excel plugins for pandas
xlrd==1.1.0
xlwt==1.3.0
# Text analysis
beautifulsoup4==4.6.0
python-Levenshtein==0.12.0
fuzzywuzzy==0.16.0
bs4==0.0.1
# Document management
num2words==0.5.6
pandocfilters==1.4.2
PyYAML==3.12
bibtexparser==1.0.1
titlecase==0.12.0
tabulate==0.8.2
# Graphing and deps - terminalplot 0.2.6
seaborn==0.8.1
matplotlib==2.2.2
backports.functools-lru-cache==1.5
cycler==0.10.0
kiwisolver==1.0.1
pyparsing==2.2.0
subprocess32==3.2.7
# Fabric
asn1crypto==0.24.0
bcrypt==3.1.4
cffi==1.11.5
cryptography==2.2.1
enum34==1.1.6
Fabric3==1.14.post1
idna==2.6
ipaddress==1.0.19
paramiko==2.4.1
pyasn1==0.4.2
pycparser==2.18
PyNaCl==1.2.1
six==1.11.0
prompter==0.3.10
# REPL - ptpython
click==6.7
click-plugins==1.0.3
cligj==0.4.0
docopt==0.6.2
jedi==0.11.1
munch==2.2.0
parso==0.1.1
ptpython==0.41
## The following requirements were added by pip freeze:
